chore: split wasm crate into submodules
